Restructuring calibration module into separate modules
